# api/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Ippo
  version: 0.0.0
tags: []
paths:
  /auth/login:
    post:
      operationId: Auth_login
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '400':
          description: 入力が不正
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 認証が必要
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /auth/logout:
    post:
      operationId: Auth_logout
      parameters: []
      responses:
        '204':
          description: There is no content to send for this request, but the headers may be useful.
  /auth/signup:
    post:
      operationId: Auth_signup
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '400':
          description: 入力が不正
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: 既に存在する
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
  /me:
    get:
      operationId: Me_get
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          description: 認証が必要
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    patch:
      operationId: Me_update
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '400':
          description: 入力が不正
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 認証が必要
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
  /posts/{postId}:
    get:
      operationId: Posts_get
      description: 投稿を1件取得する。本文は全文を返す
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: 対象が存在しない
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      operationId: Posts_remove
      description: 自分の投稿を削除する。いいねも一緒に消える
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: There is no content to send for this request, but the headers may be useful.
        '401':
          description: 認証が必要
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: 権限が無い
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象が存在しない
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /posts/{postId}/like:
    put:
      operationId: Posts_like
      description: 投稿にいいねをする
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: There is no content to send for this request, but the headers may be useful.
        '401':
          description: 認証が必要
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象が存在しない
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      operationId: Posts_unlike
      description: いいねを取り消す
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: There is no content to send for this request, but the headers may be useful.
        '401':
          description: 認証が必要
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /topics:
    get:
      operationId: Topics_list
      description: トピックの一覧を取得する
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Topic'
  /topics/{slug}/posts:
    get:
      operationId: Topics_listPosts
      description: トピックの投稿を、指定された並び順で取得する。 既定は人気順
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/SortOrder'
            default: popular
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostSummary'
        '400':
          description: 入力が不正
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象が存在しない
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      operationId: Topics_createPost
      description: トピックに投稿する
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: 入力が不正
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 認証が必要
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象が存在しない
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
components:
  schemas:
    CreatePostRequest:
      type: object
      required:
        - title
        - body
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        body:
          type: string
          minLength: 1
          maxLength: 15000
      description: 投稿の作成リクエスト
    CurrentUser:
      type: object
      required:
        - email
        - displayName
      properties:
        email:
          type: string
          format: email
          description: ログインに使うメールアドレス
        displayName:
          type: string
          description: 投稿に使われる名前
      description: ログイン中のユーザーの情報
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: ログインに使うメールアドレス
        password:
          type: string
          description: ログインに使われるパスワード
      description: ログインリクエスト
    Post:
      type: object
      required:
        - id
        - title
        - body
        - authorName
        - likeCount
        - likedByMe
        - isMine
        - createdAt
        - topic
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          description: タイトル
        body:
          type: string
          description: 本文
        authorName:
          type: string
          description: 投稿者の名前
        likeCount:
          type: integer
          format: int32
          description: いいねの数
        likedByMe:
          type: boolean
          description: 自分がいいねを押したかどうか。未ログインなら false
        isMine:
          type: boolean
          description: 自分の投稿かどうか。true のときだけ削除できる。未ログインなら false
        createdAt:
          type: string
          format: date-time
          description: 投稿日時（UTC）
        topic:
          allOf:
            - $ref: '#/components/schemas/Topic'
          description: この投稿が属するトピック。詳細ページから一覧へ戻るために使う
      description: トピックに書き込む投稿
    PostSummary:
      type: object
      required:
        - id
        - title
        - bodyPreview
        - truncated
        - authorName
        - likeCount
        - likedByMe
        - isMine
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          description: タイトル。100文字に制限しているため、一覧でも全文を返す
        bodyPreview:
          type: string
          description: 本文の先頭200文字。truncated が true なら続きがある
        truncated:
          type: boolean
          description: 本文が200文字を超えて切り詰められたかどうか
        authorName:
          type: string
          description: 投稿者の名前
        likeCount:
          type: integer
          format: int32
          description: いいねの数
        likedByMe:
          type: boolean
          description: 自分がいいねを押したかどうか。未ログインなら false
        isMine:
          type: boolean
          description: 自分の投稿かどうか。true のときだけ削除できる。未ログインなら false
        createdAt:
          type: string
          format: date-time
          description: 投稿日時（UTC）
      description: 一覧に並べる投稿。本文は先頭200文字までしか含まない
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          description: 問題の種類を表すURI。このAPIでは常に about:blank
        title:
          type: string
          description: HTTPステータスの標準の語句
        status:
          type: integer
          format: int32
          description: HTTPステータスコード
        detail:
          type: string
          description: 何が起きたかの説明
      description: RFC 9457 の Problem Details
    SignUpRequest:
      type: object
      required:
        - email
        - password
        - displayName
      properties:
        email:
          type: string
          format: email
          description: サインアップに使うメールアドレス
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: サインアップに使うパスワード
        displayName:
          type: string
          minLength: 1
          maxLength: 50
          description: サインアップに使う名前
      description: サインアップリクエスト
    SortOrder:
      type: string
      enum:
        - popular
        - newest
        - oldest
      description: ソート順
    Topic:
      type: object
      required:
        - slug
        - name
      properties:
        slug:
          type: string
          description: URLに使う識別子
        name:
          type: string
          description: 表示名
      description: 投稿を分類するカテゴリ
    UpdateProfileRequest:
      type: object
      required:
        - displayName
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 50
          description: ユーザー情報変更に使われる名前
      description: ユーザー情報変更