# backend/internal/api/gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for SortOrder.
const (
	Newest  SortOrder = "newest"
	Oldest  SortOrder = "oldest"
	Popular SortOrder = "popular"
)

// Valid indicates whether the value is a known member of the SortOrder enum.
func (e SortOrder) Valid() bool {
	switch e {
	case Newest:
		return true
	case Oldest:
		return true
	case Popular:
		return true
	default:
		return false
	}
}

// CreatePostRequest 投稿の作成リクエスト
type CreatePostRequest struct {
	Body  string `json:"body"`
	Title string `json:"title"`
}

// CurrentUser ログイン中のユーザーの情報
type CurrentUser struct {
	// DisplayName 投稿に使われる名前
	DisplayName string `json:"displayName"`

	// Email ログインに使うメールアドレス
	Email openapi_types.Email `json:"email"`
}

// LoginRequest ログインリクエスト
type LoginRequest struct {
	// Email ログインに使うメールアドレス
	Email openapi_types.Email `json:"email"`

	// Password ログインに使われるパスワード
	Password string `json:"password"`
}

// Post トピックに書き込む投稿
type Post struct {
	// AuthorName 投稿者の名前
	AuthorName string `json:"authorName"`

	// Body 本文
	Body string `json:"body"`

	// CreatedAt 投稿日時（UTC）
	CreatedAt time.Time          `json:"createdAt"`
	Id        openapi_types.UUID `json:"id"`

	// IsMine 自分の投稿かどうか。true のときだけ削除できる。未ログインなら false
	IsMine bool `json:"isMine"`

	// LikeCount いいねの数
	LikeCount int32 `json:"likeCount"`

	// LikedByMe 自分がいいねを押したかどうか。未ログインなら false
	LikedByMe bool `json:"likedByMe"`

	// Title タイトル
	Title string `json:"title"`

	// Topic この投稿が属するトピック。詳細ページから一覧へ戻るために使う
	Topic Topic `json:"topic"`
}

// PostSummary 一覧に並べる投稿。本文は先頭200文字までしか含まない
type PostSummary struct {
	// AuthorName 投稿者の名前
	AuthorName string `json:"authorName"`

	// BodyPreview 本文の先頭200文字。truncated が true なら続きがある
	BodyPreview string `json:"bodyPreview"`

	// CreatedAt 投稿日時（UTC）
	CreatedAt time.Time          `json:"createdAt"`
	Id        openapi_types.UUID `json:"id"`

	// IsMine 自分の投稿かどうか。true のときだけ削除できる。未ログインなら false
	IsMine bool `json:"isMine"`

	// LikeCount いいねの数
	LikeCount int32 `json:"likeCount"`

	// LikedByMe 自分がいいねを押したかどうか。未ログインなら false
	LikedByMe bool `json:"likedByMe"`

	// Title タイトル。100文字に制限しているため、一覧でも全文を返す
	Title string `json:"title"`

	// Truncated 本文が200文字を超えて切り詰められたかどうか
	Truncated bool `json:"truncated"`
}

// ProblemDetails RFC 9457 の Problem Details
type ProblemDetails struct {
	// Detail 何が起きたかの説明
	Detail string `json:"detail"`

	// Status HTTPステータスコード
	Status int32 `json:"status"`

	// Title HTTPステータスの標準の語句
	Title string `json:"title"`

	// Type 問題の種類を表すURI。このAPIでは常に about:blank
	Type string `json:"type"`
}

// SignUpRequest サインアップリクエスト
type SignUpRequest struct {
	// DisplayName サインアップに使う名前
	DisplayName string `json:"displayName"`

	// Email サインアップに使うメールアドレス
	Email openapi_types.Email `json:"email"`

	// Password サインアップに使うパスワード
	Password string `json:"password"`
}

// SortOrder ソート順
type SortOrder string

// Topic 投稿を分類するカテゴリ
type Topic struct {
	// Name 表示名
	Name string `json:"name"`

	// Slug URLに使う識別子
	Slug string `json:"slug"`
}

// UpdateProfileRequest ユーザー情報変更
type UpdateProfileRequest struct {
	// DisplayName ユーザー情報変更に使われる名前
	DisplayName string `json:"displayName"`
}

// TopicsListPostsParams defines parameters for TopicsListPosts.
type TopicsListPostsParams struct {
	Sort *SortOrder `form:"sort,omitempty" json:"sort,omitempty"`
}

// AuthLoginJSONRequestBody defines body for AuthLogin for application/json ContentType.
type AuthLoginJSONRequestBody = LoginRequest

// AuthSignupJSONRequestBody defines body for AuthSignup for application/json ContentType.
type AuthSignupJSONRequestBody = SignUpRequest

// MeUpdateJSONRequestBody defines body for MeUpdate for application/json ContentType.
type MeUpdateJSONRequestBody = UpdateProfileRequest

// TopicsCreatePostJSONRequestBody defines body for TopicsCreatePost for application/json ContentType.
type TopicsCreatePostJSONRequestBody = CreatePostRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /auth/login)
	AuthLogin(w http.ResponseWriter, r *http.Request)

	// (POST /auth/logout)
	AuthLogout(w http.ResponseWriter, r *http.Request)

	// (POST /auth/signup)
	AuthSignup(w http.ResponseWriter, r *http.Request)

	// (GET /me)
	MeGet(w http.ResponseWriter, r *http.Request)

	// (PATCH /me)
	MeUpdate(w http.ResponseWriter, r *http.Request)

	// (DELETE /posts/{postId})
	PostsRemove(w http.ResponseWriter, r *http.Request, postId string)

	// (GET /posts/{postId})
	PostsGet(w http.ResponseWriter, r *http.Request, postId string)

	// (DELETE /posts/{postId}/like)
	PostsUnlike(w http.ResponseWriter, r *http.Request, postId string)

	// (PUT /posts/{postId}/like)
	PostsLike(w http.ResponseWriter, r *http.Request, postId string)

	// (GET /topics)
	TopicsList(w http.ResponseWriter, r *http.Request)

	// (GET /topics/{slug}/posts)
	TopicsListPosts(w http.ResponseWriter, r *http.Request, slug string, params TopicsListPostsParams)

	// (POST /topics/{slug}/posts)
	TopicsCreatePost(w http.ResponseWriter, r *http.Request, slug string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// AuthLogin operation middleware
func (siw *ServerInterfaceWrapper) AuthLogin(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AuthLogin(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AuthLogout operation middleware
func (siw *ServerInterfaceWrapper) AuthLogout(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AuthLogout(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AuthSignup operation middleware
func (siw *ServerInterfaceWrapper) AuthSignup(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AuthSignup(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// MeGet operation middleware
func (siw *ServerInterfaceWrapper) MeGet(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.MeGet(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// MeUpdate operation middleware
func (siw *ServerInterfaceWrapper) MeUpdate(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.MeUpdate(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostsRemove operation middleware
func (siw *ServerInterfaceWrapper) PostsRemove(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "postId" -------------
	var postId string

	err = runtime.BindStyledParameterWithOptions("simple", "postId", r.PathValue("postId"), &postId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "postId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostsRemove(w, r, postId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostsGet operation middleware
func (siw *ServerInterfaceWrapper) PostsGet(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "postId" -------------
	var postId string

	err = runtime.BindStyledParameterWithOptions("simple", "postId", r.PathValue("postId"), &postId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "postId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostsGet(w, r, postId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostsUnlike operation middleware
func (siw *ServerInterfaceWrapper) PostsUnlike(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "postId" -------------
	var postId string

	err = runtime.BindStyledParameterWithOptions("simple", "postId", r.PathValue("postId"), &postId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "postId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostsUnlike(w, r, postId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostsLike operation middleware
func (siw *ServerInterfaceWrapper) PostsLike(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "postId" -------------
	var postId string

	err = runtime.BindStyledParameterWithOptions("simple", "postId", r.PathValue("postId"), &postId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "postId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostsLike(w, r, postId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// TopicsList operation middleware
func (siw *ServerInterfaceWrapper) TopicsList(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.TopicsList(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// TopicsListPosts operation middleware
func (siw *ServerInterfaceWrapper) TopicsListPosts(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameterWithOptions("simple", "slug", r.PathValue("slug"), &slug, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "slug", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params TopicsListPostsParams

	// ------------- Optional query parameter "sort" -------------

	err = runtime.BindQueryParameterWithOptions("form", false, false, "sort", r.URL.Query(), &params.Sort, runtime.BindQueryParameterOptions{Type: "", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "sort"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "sort", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.TopicsListPosts(w, r, slug, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// TopicsCreatePost operation middleware
func (siw *ServerInterfaceWrapper) TopicsCreatePost(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameterWithOptions("simple", "slug", r.PathValue("slug"), &slug, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "slug", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.TopicsCreatePost(w, r, slug)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/auth/login", wrapper.AuthLogin)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/auth/logout", wrapper.AuthLogout)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/auth/signup", wrapper.AuthSignup)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/me", wrapper.MeGet)
	m.HandleFunc(http.MethodPatch+" "+options.BaseURL+"/me", wrapper.MeUpdate)
	m.HandleFunc(http.MethodDelete+" "+options.BaseURL+"/posts/{postId}", wrapper.PostsRemove)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/posts/{postId}", wrapper.PostsGet)
	m.HandleFunc(http.MethodDelete+" "+options.BaseURL+"/posts/{postId}/like", wrapper.PostsUnlike)
	m.HandleFunc(http.MethodPut+" "+options.BaseURL+"/posts/{postId}/like", wrapper.PostsLike)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/topics", wrapper.TopicsList)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/topics/{slug}/posts", wrapper.TopicsListPosts)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/topics/{slug}/posts", wrapper.TopicsCreatePost)

	return m
}